Provision XMPP user for Jigasi
The Gateway to SIP functionality uses an XMPP client connection to represent SIP-based participants in a meeting.
Such client connections can use anonymous authentication, but that's not always enabled in Openfire.

This commit adds functionality that allows an administrator to provision an XMPP account for the Gateway to SIP
functionality. It attempts to automatically generate an account if the User Provider allows for this.